requirements.txt
openai
scikit-learn
joblib
langchain-community
